+++ packages/helux-core/src/hooks/index.ts
@@ -10,3 +10,4 @@ export { useReactive, useReactiveX } from './useReactive';
 export { storeSrv, useService } from './useService';
 export { useWatch, useWatchEffect } from './useWatch';
 export { useLockDep } from './useLockDep';
+export { useObject } from './useObject';