+++ packages/helux-hooks-impl/src/api.ts
@@ -1,7 +1,7 @@
 import { useEffect, useLayoutEffect } from './hooks/useEffect';
 import { useForceUpdate } from './hooks/useForceUpdate';
 import { useIsStrict } from './hooks/useIsStrict';
-import { useObject, useObjectLogic } from './hooks/useObject';
+import { useObject, useObjectLogic, useObjectLogicV2 } from './hooks/useObject';
 import { useStable } from './hooks/useStable';
 
-export { useIsStrict, useObject, useObjectLogic, useForceUpdate, useEffect, useLayoutEffect, useStable };
+export { useIsStrict, useObject, useObjectLogic, useObjectLogicV2, useForceUpdate, useEffect, useLayoutEffect, useStable };