+++ packages/helux/src/index.ts
@@ -1,4 +1,5 @@
 import { initHeluxContext } from '@helux/core';
+import type { HooksApiImpl } from '@helux/hooks-impl';
 import * as React from 'react';
 
 const api = initHeluxContext({ heluxCtxKey: '__HELUX__', reactLib: React });
@@ -8,6 +9,11 @@ const api = initHeluxContext({ heluxCtxKey: '__HELUX__', reactLib: React });
  * 这个接口仅为了兼容 helux v2 升级到 v3 后不报错
  */
 export const createShared = api.share;
+
+/**
+ * 避免警告：如果没有引用 "../hooks-impl/src/types-api"，则无法命名 "hookImpl" 的推断类型。这很可能不可移植。需要类型注释
+ */
+export const hookImpl: HooksApiImpl = api.hookImpl;
 
 // 导出 core 所有方法，类型由 index.d.ts 提供（见 package.json 的 types 配置）
 export const {