+++ packages/helux-hooks-impl/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @helux/hooks-impl
 
+## 5.4.6
+
+### Patch Changes
+
+- f92ad32: mark type HooksApiImpl
+  - @helux/types@5.4.6
+  - @helux/utils@5.4.6
+
+## 5.4.4
+
+### Patch Changes
+
+- 0e4bf25: expose hookImpl, add useObject to helux-core
+  - @helux/types@5.4.4
+  - @helux/utils@5.4.4
+
 ## 5.4.3
 
 ### Patch Changes