+++ packages/helux-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @helux/utils
 
+## 5.4.6
+
+### Patch Changes
+
+- @helux/types@5.4.6
+
+## 5.4.4
+
+### Patch Changes
+
+- @helux/types@5.4.4
+
 ## 5.4.3
 
 ### Patch Changes